==> R/91_descriptives_general.Rmd <==
---
title: "Luxembourg Research Evaluation 2022: General Research Field Descrition"
author: "Daniel S. Hain"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    theme: flatly
    code_folding: hide
    df_print: paged
    toc: false
    toc_depth: 2
    toc_float:
      collapsed: false
params:
    institute: 
       value: null
    department:
       value: null
---

<!---
# Add to YAML when reviewing
  html_notebook:
    theme: flatly
    code_folding: hide
    df_print: paged
    toc: false
    toc_depth: 2
    toc_float:
      collapsed: false
--->

```{=html}
<style type="text/css">
.main-container {
  max-width: 1200px;
  margin-left: auto;
  margin-right: auto;
}
</style>
```

```{r setup, include=FALSE}
### Generic preamble
#rm(list=ls())
Sys.setenv(LANG = "en")
options(scipen = 5)
set.seed(1337)

### Load packages  
# general
library(tidyverse)
library(magrittr)

# Kiblio & NW
library(bibliometrix)
library(tidygraph)
library(ggraph)

# NLP
library(tidytext)

# Knit
library(knitr) # For display of the markdown
library(kableExtra) # For table styling

# own functions
source("../functions/functions_basic.R")
source("../functions/functions_summary.R")
source("../functions/00_parameters.R")

# Knitr options
knitr::opts_chunk$set(echo = FALSE, 
                      warning = FALSE, 
                      message = FALSE)
```

<!-- ####################################################################################### -->
<!-- ####################################################################################### -->
<!-- ############################# NEXT PART ############################################### -->
<!-- ####################################################################################### -->
<!-- ####################################################################################### -->

```{r, include=FALSE}
#var_inst <- 'LISER'
#var_dept <- 'UD'
```

```{r}
var_inst <- params$institute
var_dept <- params$department
```

# Introduction: `r var_inst` Department `r var_dept`

```{r, include=FALSE}
# Load data
M <- readRDS(paste0('../../temp/M_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '.rds')) %>% as_tibble() %>% 
  distinct(UT, .keep_all = TRUE) %>% 
  filter(PY >= PY_min, PY <= PY_max)
```

Here are preliminary results of the bibliometric mapping of the 2022 Luxembourg research evaluation. Its purpose is:

* To map the broader research community and distinct research field the department contributes to.
* Identify core knowledge bases, research areaS, TRENDS AND TOPICS.
* Highlight the positioning of the department within this dynamics.

The method for the research-field-mapping can be reiviewed here:

[Rakas, M., & Hain, D. S. (2019). The state of innovation system research: What happens beneath the surface?. Research Policy, 48(9), 103787.](https://doi.org/10.1016/j.respol.2019.04.011)



# Seed Articles

```{r, include=FALSE}
seed <-convert2df(file = paste0('../../data/seeds/scopus_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '_seed_select.csv'), dbsource = "scopus", format = "csv") %>%
  as_tibble() %>%
  mutate(seed = TRUE) 
```

The seed articles deemed representative for the active areas of research in the institution, and include authors affiliated with the institution. They can be selected in three ways:

1. Via bibliographic clustering of the institutions publications and selection of most central articles per cluster (only clsuters where n >= 0.05N). Selection can be found at `r paste0('https://github.com/daniel-hain/biblio_lux_2022/blob/master/output/seed/scopus_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '_seed.csv')`
2. Manual selection of relevant publications.
3. A combination of 1. and 2.

The present analysis is based on the following seed articles:

```{r}
seed %>%
  select(AU, PY, TI, JI) %>%
  mutate(AU = AU %>% str_trunc(30),
         TI = TI %>% str_trunc(100),
         JI = JI %>% str_trunc(30)) %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), font_size = 10)
```

# General Overview over articles {.tabset}

## Main Indicators: Publications, Authors, Countries (Tabular)

```{r, include=FALSE}
# results <- biblioAnalysis(M, sep = ";")
results <- read_rds(paste0('../../temp/M_res_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '.rds'))
```

```{r}
# RODO figure out how to not plot all
results %>% summary(k = 10, pause = FALSE)
```

## Main Indicators: Publications, Authors, Countries (Graphical)

```{r}
results %>% plot(k = 10, pause = FALSE)
```

```{r}
prod_AU <- M %>% authorProdOverTime(k = 10, graph = TRUE)
#plot(prod_AU$graph)
```

```{r, include=FALSE}
rm(results, prod_AU)
```

# Cited references

```{r, include=FALSE}
CR <- readRDS(paste0('../../temp/CR_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '.rds')) 
```

Top 20 cited references (by corpus documents):

```{r}
CR$Cited %>% as_tibble() %>% head(20) %>% 
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "hover", "condensed", "responsive"), font_size = 10)
```

```{r, include=FALSE}
rm(CR)
```

<!-- ####################################################################################### -->
<!-- ####################################################################################### -->
<!-- ############################# NEXT PART ############################################### -->
<!-- ####################################################################################### -->
<!-- ####################################################################################### -->

# Additional analysis

## Authors, Themes & Journals

```{r, include=FALSE}
M_threefield <- readRDS(paste0('../../temp/threefield_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '.rds')) 
```

```{r, fig.width=17.5, fig.height=17.5}
M_threefield
```

```{r}
#rm(M_threefield)
```

```{r, fig.width=17.5, fig.height=17.5}
### Conceptual trajectories: Historical citation path analysis
#histResults <- readRDS(paste0('../temp/histResult_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '.rds')) 
#histResults %>% histPlot(n =50, size = 10, labelsize = 7.5)
#rm(histResults)
```

# Endnotes

All results are preliminary so far...

```{r}
# After knitted do this
#file.rename(from = "91_descriptives_general.nb.html", to = paste0('../output/field_mapping/field_mapping_general_', str_to_lower(var_inst), '_', str_to_lower(var_dept), '.html'))
```
